+++ src/mock/const.js
@@ -1,0 +1,21 @@
+const MAX_SIZE_DESCRIPTION = 5;
+
+const COUNT_PHOTO = 50;
+
+const MAX_SIZE_PHOTO = 6;
+
+const DESCRIOTIONS = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ',
+  'Cras aliquet varius magna, non porta ligula feugiat eget. ',
+  'Fusce tristique felis at fermentum pharetra. ',
+  'Aliquam id orci ut lectus varius viverra. ',
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. ',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. ',
+  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. ',
+  'Sed sed nisi sed augue convallis suscipit in sed felis. ',
+  'Aliquam erat volutpat. ',
+  'Nunc fermentum tortor ac porta dapibus. ',
+  'In rutrum ac purus sit amet tempus. ',
+];
+
+export { MAX_SIZE_PHOTO, MAX_SIZE_DESCRIPTION, COUNT_PHOTO, DESCRIOTIONS };
